minors
git-svn-id: http://svn.geogr.uni-jena.de/svn/jamsmodels/trunk@1939 503a54e1-1403-0410-a754-803ee147b016

+++ J2K_base/src/org/unijena/j2k/regionalisation/IDWWeightCalculator.java
@@ -114,9 +114,9 @@ public class IDWWeightCalculator extends JAMSComponent {
 
         boolean isLatLon = latLong != null && latLong.getValue();
         if (isLatLon) {
-            idw.init(statX, statY, null, (int) pidw.getValue(), IDW.Projection.LATLON);
+            idw.init(statX, statY, null, pidw.getValue(), IDW.Projection.LATLON);
         } else {
-            idw.init(statX, statY, null, (int) pidw.getValue(), IDW.Projection.ANY);
+            idw.init(statX, statY, null, pidw.getValue(), IDW.Projection.ANY);
         }
     }
